Add Unescaper and TBXMLParserDelegate
I've moved the Unescaping method to a new object so it can use it can be
used with more than one parser. Noticed this unescaper isn't full proof
and needs adding to.

Implemented the TBSMLparser you can swap between them in the demo. I
have found a bug in the TBXMLparser, when an element can't be found it
silently fails instead of raising an exception.Which might scupper a
one size fits all RSS feed parser. But hell it's open source so people
can tailor it!